Templates moved - initial scroll snap imp
