--- web/src/pages/index.module.scss
@@ -1,3 +1,0 @@
-.listWrapper {
-	display: flex;
-}
