+++ web/src/templates/FoodPage.module.scss
@@ -1,0 +1,34 @@
+.container {
+	position: relative;
+	display: flex;
+}
+
+.listWrapper {
+	padding: 1.2rem;
+	scroll-snap-align: start;
+}
+
+.spotList {
+	display: none;
+}
+
+.isSpotPage {
+	.postList {
+		// padding-right: 14.6rem;
+	}
+}
+
+@media all and (min-width: 30rem) {
+	.container {
+		width: 100vw;
+		padding: 0 2.4rem;
+		transform: translate3d(-2.4rem, 0, 0);
+		overflow-x: scroll;
+		scroll-snap-type: x mandatory;
+		scroll-padding-left: 2.4rem;
+	}
+
+	.spotList {
+		display: block;
+	}
+}
