+++ web/src/templates/FoodPostPage.js
@@ -1,15 +1,15 @@
+import styles from './FoodPage.module.scss';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { graphql } from 'gatsby';
-import Container from '../../components/layout/Container';
-import PostList from '../../components/Spots/PostList/PostList';
+import Container from '../components/layout/Container';
+import PostList from '../components/Spots/PostList/PostList';
 import PostDetail from './PostDetail/PostDetail';
-import Header from '../../components/layout/Header/Header';
-import { replaceAllNonCharacters } from '../../utils/utils';
-import SpotList from '../../components/Spots/SpotList/SpotList';
+import Header from '../components/layout/Header/Header';
+import { replaceAllNonCharacters } from '../utils/utils';
+import SpotList from '../components/Spots/SpotList/SpotList';
 
 export const query = graphql`
-	# query foodPostPageQuery($foodSpotId: String!, $foodPostId: String!) {
 	query foodPostPageQuery($foodPostId: String!, $foodSpotId: String!) {
 		spots: allSanityFoodSpot {
 			totalCount
@@ -60,11 +60,11 @@ const FoodPostPage = ({
 	return (
 		<Container hasBackButton backURL={`/foodspot/${spotSlug}`}>
 			<Header title={foodPost.title} subTitle={name} />
-			<div style={{ display: 'flex' }}>
-				<div style={{ padding: '1.2rem' }}>
+			<div className={`${styles.container}`}>
+				<div className={`${styles.listWrapper} ${styles.spotList}`}>
 					<SpotList spots={spots.edges} />
 				</div>
-				<div style={{ padding: '1.2rem' }}>
+				<div className={`${styles.listWrapper}`}>
 					<PostDetail
 						post={foodPost}
 						posts={filteredPosts}
