+++ web/src/templates/FoodSpotPage.js
@@ -1,11 +1,12 @@
+import styles from './FoodPage.module.scss';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { graphql } from 'gatsby';
-import Container from '../../components/layout/Container';
-import Header from '../../components/layout/Header/Header';
-import { OnePageContextConsumer } from '../../context/OnePageContext';
-import SpotList from '../../components/Spots/SpotList/SpotList';
-import PostList from '../../components/Spots/PostList/PostList';
+import Container from '../components/layout/Container';
+import Header from '../components/layout/Header/Header';
+import { OnePageContextConsumer } from '../context/OnePageContext';
+import SpotList from '../components/Spots/SpotList/SpotList';
+import PostList from '../components/Spots/PostList/PostList';
 
 export const query = graphql`
 	query foodSpotPageQuery($foodSpotId: String!) {
@@ -52,11 +53,15 @@ const FoodSpotPage = props => {
 				return (
 					<Container>
 						<Header title={name} subTitle="Visits" />
-						<div style={{ display: 'flex' }}>
-							<div style={{ padding: '1.2rem' }}>
+						<div className={`${styles.container} ${styles.isSpotPage}`}>
+							<div
+								className={`${styles.listWrapper} ${styles.spotList}`}
+							>
 								<SpotList spots={spots.edges} />
 							</div>
-							<div style={{ padding: '1.2rem' }}>
+							<div
+								className={`${styles.listWrapper} ${styles.postList}`}
+							>
 								<PostList
 									posts={posts}
 									currentPath={pathContext.currentPath}
